views: make params parameter for view_docs{,_with_list} optional
The documentation says that `params` argument of the db.view() and db.view_with_list()
functions is optional. If omitting the parameter instead of passing an empty object, it
complains about a "TypeError: Cannot read property `keys` of undefined".

With this change, it is possible to omit the `params` argument without the exception.

Signed-off-by: Nuno Job <nunojobpinto@gmail.com>