Add further documentation for the db.insert function
The db.insert function can be used to both create and update documents. This wasn't clear from the the README and we had more than one issue raise about it on our Cloudant fork of the project. This commit shows using the function with one parameter and a callaback, and its use to update existing documents.